Rename components to be a consistent name format "ComponentNameHere.vue`
