+++ src/components/ChatHeader.vue
